Separate groq_llm_cost.py
