+++ council/llm/providers/groq/groq_llm.py
@@ -1,22 +1,11 @@
 from __future__ import annotations
 
-from typing import Any, Dict, List, Mapping, Optional, Sequence
+from typing import Any, List, Sequence
 
 from council.contexts import Consumption, LLMContext
-from council.llm import (
-    DefaultLLMConsumptionCalculatorHelper,
-    LLMBase,
-    LLMConfigObject,
-    LLMCostCard,
-    LLMCostManagerObject,
-    LLMMessage,
-    LLMMessageRole,
-    LLMProviders,
-    LLMResult,
-)
+from council.llm import LLMBase, LLMConfigObject, LLMMessage, LLMMessageRole, LLMProviders, LLMResult
 from council.utils.utils import DurationManager
 from groq import Groq
-from groq.types import CompletionUsage
 from groq.types.chat import (
     ChatCompletionAssistantMessageParam,
     ChatCompletionMessageParam,
@@ -26,45 +15,7 @@ from groq.types.chat import (
 from groq.types.chat.chat_completion import ChatCompletion, Choice
 
 from .groq_llm_configuration import GroqLLMConfiguration
-
-
-class GroqConsumptionCalculator(DefaultLLMConsumptionCalculatorHelper):
-    _cost_manager = LLMCostManagerObject.groq()
-    COSTS: Mapping[str, LLMCostCard] = _cost_manager.get_cost_map("default")
-
-    def __init__(self, model: str) -> None:
-        super().__init__(model)
-
-    def get_consumptions(self, duration: float, usage: Optional[CompletionUsage]) -> List[Consumption]:
-        if usage is None:
-            return self.get_default_consumptions(duration)
-
-        prompt_tokens = usage.prompt_tokens
-        completion_tokens = usage.completion_tokens
-        return (
-            self.get_base_consumptions(duration, prompt_tokens=prompt_tokens, completion_tokens=completion_tokens)
-            + self.get_duration_consumptions(usage)
-            + self.get_cost_consumptions(prompt_tokens=prompt_tokens, completion_tokens=completion_tokens)
-        )
-
-    def get_duration_consumptions(self, usage: CompletionUsage) -> List[Consumption]:
-        """Optional duration consumptions specific to Groq."""
-        usage_times: Dict[str, Optional[float]] = {
-            "queue_time": usage.queue_time,
-            "prompt_time": usage.prompt_time,
-            "completion_time": usage.completion_time,
-            "total_time": usage.total_time,
-        }
-
-        consumptions = []
-        for key, value in usage_times.items():
-            if value is not None:
-                consumptions.append(Consumption.duration(value, f"{self.model}:groq_{key}"))
-
-        return consumptions
-
-    def find_model_costs(self) -> Optional[LLMCostCard]:
-        return self.COSTS.get(self.model)
+from .groq_llm_cost import GroqConsumptionCalculator
 
 
 class GroqLLM(LLMBase[GroqLLMConfiguration]):
